refactor: set up bookings feature and centralized routes
