+++ src/pages/admin/BookingsPage.tsx
@@ -1,6 +1,6 @@
 
 import React from 'react';
-import BookingsManagement from '@/components/bookings/BookingsManagement';
+import BookingsManagement from '@/features/bookings/components/BookingsManagement';
 
 const BookingsPage = () => {
   return <BookingsManagement />;
